Improve the API and add additional documentation

+++ src/main/java/com/martin/aleksandrov/EVChargingStationAPI/services/impl/ChargingStationServiceImpl.java
@@ -1,5 +1,6 @@
 package com.martin.aleksandrov.EVChargingStationAPI.services.impl;
 
+import com.martin.aleksandrov.EVChargingStationAPI.exceptions.ChargingStationNotFoundException;
 import com.martin.aleksandrov.EVChargingStationAPI.models.dtos.ChargingStationCreateDto;
 import com.martin.aleksandrov.EVChargingStationAPI.models.dtos.ChargingStationDto;
 import com.martin.aleksandrov.EVChargingStationAPI.models.entities.ChargingStationEntity;
@@ -8,9 +9,13 @@ import com.martin.aleksandrov.EVChargingStationAPI.services.ChargingStationServi
 import lombok.AllArgsConstructor;
 import org.apache.coyote.BadRequestException;
 import org.modelmapper.ModelMapper;
+import org.springframework.boot.context.properties.source.InvalidConfigurationPropertyValueException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.MissingResourceException;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -82,8 +87,16 @@ public class ChargingStationServiceImpl implements ChargingStationService {
 
 
     @Override
+    @Transactional
     public void deleteChargingStation(String uniqueId) {
-        this.chargingStationRepository.deleteByUniqueId(uniqueId);
-    }
+        Optional<ChargingStationEntity> byUniqueId = this.chargingStationRepository.findChargingStationEntityByUniqueId(uniqueId);
+        if (byUniqueId.isPresent()) {
 
+           this.chargingStationRepository.deleteByUniqueId(uniqueId);
+        } else {
+
+            throw new ChargingStationNotFoundException("Missing charging station with such id");
+        }
+
+    }
 }
